chore: further organize functional tests and fixtures
