+++ tests/integration/src/rust_pipeline/functional/abi/enums.rs
@@ -1,0 +1,9 @@
+use crate::CompilerTest;
+
+#[test]
+fn enum_matching() {
+    let _ = CompilerTest::rust_source_program(include_str!(
+        "../../../../../fixtures/standalone/enum.rs"
+    ))
+    .compile_package();
+}
