+++ tests/integration/src/rust_pipeline/functional/abi/mod.rs
@@ -1,0 +1,4 @@
+mod arrays;
+mod enums;
+mod functions;
+mod transforms;
