+++ tests/integration/src/rust_pipeline/functional/abi/transforms/tx_kernel/get_inputs.rs
