--- tests/integration/src/rust_pipeline/functional/apps/mod.rs
@@ -1,4 +1,0 @@
-mod felt_intrinsics;
-mod function_call_hir2;
-mod mem_intrinsics_heap_base;
-mod support;
