+++ tests/integration/src/rust_pipeline/functional/debuginfo/mod.rs
@@ -1,0 +1,1 @@
+mod source_locations;
