+++ tests/integration/src/rust_pipeline/functional/debuginfo/source_locations.rs
