+++ tests/integration/src/rust_pipeline/functional/intrinsics/felt.rs
@@ -6,16 +6,16 @@ use midenc_frontend_wasm::WasmTranslationConfig;
 use midenc_hir::Felt;
 use proptest::{prelude::*, test_runner::TestRunner};
 
-use super::support::cargo_toml;
 use crate::{
     CompilerTest, CompilerTestBuilder,
     cargo_proj::project,
     compiler_test::{sdk_alloc_crate_path, sdk_crate_path},
+    rust_pipeline::functional::support::cargo_toml,
 };
 
 #[test]
-fn mem_intrinsics_heap_base() {
-    let name = "mem_intrinsics_heap_base";
+fn basic_felt_arithmetic() {
+    let name = "felt_intrinsics";
     let cargo_proj = project(name)
         .file("Cargo.toml", &cargo_toml(name))
         .file(
@@ -23,10 +23,6 @@ fn mem_intrinsics_heap_base() {
             r#"
                 #![no_std]
                 #![feature(alloc_error_handler)]
-
-                // Global allocator to use heap memory in no-std environment
-                #[global_allocator]
-                static ALLOC: miden_sdk_alloc::BumpAlloc = miden_sdk_alloc::BumpAlloc::new();
 
                 // Required for no-std crates
                 #[panic_handler]
@@ -40,12 +36,15 @@ fn mem_intrinsics_heap_base() {
                     loop {}
                 }
 
-                extern crate alloc;
-                use alloc::{vec, vec::Vec};
+                // Global allocator to use heap memory in no-std environment
+                #[global_allocator]
+                static ALLOC: miden::BumpAlloc = miden::BumpAlloc::new();
+
+                use miden::*;
 
                 #[unsafe(no_mangle)]
-                pub fn entrypoint(a: u32) -> Vec<u32> {
-                    vec![a*2]
+                pub fn entrypoint(a: Felt, b: Felt) -> Felt {
+                   a / (a * b - a + b)
                 }
             "#,
         )
