+++ tests/integration/src/rust_pipeline/functional/intrinsics/heap_base.rs
@@ -6,16 +6,16 @@ use midenc_frontend_wasm::WasmTranslationConfig;
 use midenc_hir::Felt;
 use proptest::{prelude::*, test_runner::TestRunner};
 
-use super::support::cargo_toml;
 use crate::{
     CompilerTest, CompilerTestBuilder,
     cargo_proj::project,
     compiler_test::{sdk_alloc_crate_path, sdk_crate_path},
+    rust_pipeline::functional::support::cargo_toml,
 };
 
 #[test]
-fn function_call_hir2() {
-    let name = "function_call_hir2";
+fn heap_base() {
+    let name = "mem_intrinsics_heap_base";
     let cargo_proj = project(name)
         .file("Cargo.toml", &cargo_toml(name))
         .file(
@@ -25,8 +25,8 @@ fn function_call_hir2() {
                 #![feature(alloc_error_handler)]
 
                 // Global allocator to use heap memory in no-std environment
-                // #[global_allocator]
-                // static ALLOC: miden::BumpAlloc = miden::BumpAlloc::new();
+                #[global_allocator]
+                static ALLOC: miden_sdk_alloc::BumpAlloc = miden_sdk_alloc::BumpAlloc::new();
 
                 // Required for no-std crates
                 #[panic_handler]
@@ -40,17 +40,12 @@ fn function_call_hir2() {
                     loop {}
                 }
 
-                // use miden::Felt;
+                extern crate alloc;
+                use alloc::{vec, vec::Vec};
 
                 #[unsafe(no_mangle)]
-                #[inline(never)]
-                pub fn add(a: u32, b: u32) -> u32 {
-                    a + b
-                }
-
-                #[unsafe(no_mangle)]
-                pub fn entrypoint(a: u32, b: u32) -> u32 {
-                    add(a, b)
+                pub fn entrypoint(a: u32) -> Vec<u32> {
+                    vec![a*2]
                 }
             "#,
         )
