+++ tests/integration/src/rust_pipeline/functional/intrinsics/mod.rs
@@ -1,0 +1,2 @@
+mod felt;
+mod heap_base;
