+++ tests/integration/src/rust_pipeline/functional/memory/global_variables.rs
@@ -1,0 +1,9 @@
+use crate::CompilerTest;
+
+#[test]
+fn global_variable_static_mut() {
+    let _ = CompilerTest::rust_source_program(include_str!(
+        "../../../../../fixtures/standalone/static_mut.rs"
+    ))
+    .compile_package();
+}
