+++ tests/integration/src/rust_pipeline/functional/support.rs
