--- tests/integration/src/rust_pipeline/functional/types/mod.rs
@@ -1,3 +1,0 @@
-mod array;
-mod enum_type;
-mod static_mut;
