--- tests/integration/src/rust_pipeline/functional/types/static_mut.rs
@@ -1,7 +1,0 @@
-use crate::CompilerTest;
-
-#[test]
-fn test_static_mut() {
-    let _ = CompilerTest::rust_source_program(include_str!("types_src/static_mut.rs"))
-        .compile_package();
-}
